remove trail content for all stories on email except for the head
